Upated plotSpec's handling of vertical bounds to handle cases where only one or none are gven.
The old behavior when no bounds were given was to fix the lower bound at zero
while setting the upper bound to 1.15 times the maximum flux found in the region.
This can now be replicated by passing only a lower bound of zero. The new
behavior when passing no vertical bounds sets the upper bound to 1.15 times the
maximum flux and the lower bound to 0.85 times the minimum flux found in the
region being plotted.