obs2d.py: Fixed up handling of gain-correction.
getPhotonFluxArray and getErrorArray now both work off the new
self._gainCorrectedFluxArray property, which is self._rawFluxArray corrected
with the gain from the header. Both of them can create it if it doesn't yet
exist, but it should only be created once since both of them will likely be
called at the same time.

This should help make sure that the error array is handled correctly, and
is always constructed from the photon array before the photon array is blaze
corrected. The error array can then be blaze corrected, same as the photon
array.