measure_spectral_RMS: Major restructuring.
Removed the '--ADP' and '--e2ds' flags (and associated code for plotting
ADP files, restructured the '--update' flag to accept multiple arguments for
individual HDUs to update in the file, and added a '--names' flag that also
accepts multiple arguments of star names (or 'Synthetic') to analyze. Code
now automatically chooses the correct base directory if 'Synthetic' is given
as a star name or not.

Reworked the graphs to show normalized RMS and median errors, and to plot
the one-to-one line based on the x-limits of the sample for better automatic
visibility.