transition_line: Added __eq__ method to the Transition class, updated tests.
The __eq__ method will progressively check various attributes to see if they
match between the two things to be compared. If any of them don't match, it
will return False. If it gets to the end without managing to prove the two are
different, it will return True.

Currently the list of attributes checked are: wavelength, atomicNumber,
ionizationState, lowerEnergy, higherEnergy, lowerJ, higherJ, lowerOrbital,
and higherOrbital. Since only the first three of those are guaranteed to
exist in a Transition instance, it will first check that the other are present
in both instances before comparing them.

If the second thing to be compared is not an instance of Transition, __eq__
will return False.