Added plots of information, better matching of transitions and pairs.
TransitionLine and TransitionPair objects now have a 'label' property, a
string which contains some basic information about them (wavelength [to
seven significant digits], element, and ionization state for TransitionLine,
and the labels of both constituent TransitionLines separated by an underscore
for TransitionPairs). This string should be unique to a transition and pair
of transitions, and hopefully fully deterministic, making matching transitions
and pairs a breeze.

I also added plotting functionality to the script (controlled by the --create-
plots flag), which will create a plot for each pair containing all the
measurements of that pair found in the observations of the star that was
passed to the script.