Check shape of data array before storing, and transpose if necessary.
At some point it seems that the default shape of arrays imported with astropy changed, so that HARPS e2ds files went from (72, 4096) to (4096, 72). This broke a lot of code, so upon importing data from a file __init__ first checks its shape and transposes it if necessary.

Also in the test file, the 'use_new_coefficients' and 'use_pixel_positions' keywords are explicitly called.