Added check to reshape raw data if necessary, fix for bugs with calling pixel edge arrays.
The data arrays in the pixel geometry files from C. Lovis are opposite normal
HARPS orientation, being (4096, 72) instead of (72, 4096). Previously in
commit b09491067c36913e7478b87adc533426eb100e51 I checked for this and
transposed the arrays, but that turned out to be very much the wrong thing to
do. This commit adds a new method to HARPSFile2D called _reshape_if_necessary
which checks the shape of the input array (the raw data array from a HARPS
file) and reshapes it to be (72, 4096) if it is (4096, 72). This correctly
fixes the problem, and creates an array correctly laid out.

There was also a bug with calling pixel edge arrays, which should be fixed
now.

Also fixed a crash from tqdm.write trying to write out a Path by wrapping it
in str().