Added 'airmass' property, improved test coverage.
The airmass is now taken from the header cards which detail the
start and end airmass for each observation. (The airmass is
calculated as the mean of the two.)

Removed some extraneous try blocks around header cards which really
should be in every HARPS file (if they're not, something has gone
drastically wrong and the read will fail anyway).

Added a test which involves updating an already-opened file, which
improved test coverage quite a bit.