Added new CLI arguments.
Added a flag '--create-fit-plots', which replicates the creation of
the plots showing the fits to absorption features that is done in
find_transitions.py. This is intended to allow rapid updating of
those plots should they need it without needing to rerun the entire
fitting process again. However, they are still created by default
when running find_transitions.py, as it doesn't increase the overall
running time by that much and there's no particular reason not to;
this option simply allows recreating them if necessary without fitting
again.

Added a flag '--use-tex' which toggles the use of TeX rendering for
text on -- note that this *drastically* slows down the rendering for
the fit plots, so should be used with care. The pair plots require it
to be active to work, so the flag '--create-plots' will also silently
enable TeX rendering.