Added functionality to work out pairs based on transition locations on the HARPS CCD, and include this in the pickled lists.
Added CLI flags '--incorporate-position-information' to do this, and
'-v, --verbose' for printing additional information.

Added functions 'save_transitions_and_pairs' and
'format_transitions_and_pairs' to handle saving pickled lists and
formatting text files of the final selection of transitions and pairs.

The --incorporate-position-information flag will now look at where
transitions fall on the CCD, determine which orders they need to be
fitted in in order for each pair they're a part of to have both
transitions measured on the same order, then store that information
as the attribute `ordersToMeasureIn` for TransitionPair objects and as
`ordersToFitIn` for Transition objects.