Massive refactor to allow for saving and loading data
Previously initializing a Star would require reading all the
pickled fit files each time, and was massively inefficient. Now,
the accumulated data is automatically dumped upon successful
initialization, and the presence of such a data file is checked
for before reading the pickled fit files. If one is found, the
instance has various attributes filled out from the HDF5 file.

Added transitions_list and pairs_list properties (with setter
methods) to handle getting the transition and pair lists.

Added _dumpDataToDisk and constructFromHDF5 methods to handle
saving and loading collected data in HDF5 files. The data stored,
currently, are the arrays of fit means and errors, the arrays of
pair separations and errors, and dictionaries of transition and
pair labels and observation dates.