Added multiple plot types and arguments to select them.
Added get_star_results function to generate results consistently from stars for making different plots. It returns a dictionary of various information in case a plot doesn't need all of it.

Made the code to plot the chi-squared distribution its own function to be able to plot it for multiple plot types.