Added initial support for creating Star from observations with missing data.
Specifically, when constructing a Star from a directory where failed fits are specified with None in the pickled list of fits, the resulting arrays will have 'nan's inserted as appropriate. This also propagates to the pair separation arrays.

Also updated tests to check BERV, and corrected incorrect usage of 'pytest.approx' in a few tests which was causing them to always evaluate True in assert statements.