Fix for case where first fit was None.
Trying to create a star where an observation's list of fits had failed on the first entry would fail, as the first fit was used for getting the observation date, BERV, and airmass for the observation. The program will now iterate through the list until it find a non-None fit to use. (If *all* the fits are None it should crash later on when trying to add units to self.bervArray.)