Added verbose_print function.
verbose_print returns a different function depending on whether it is passed a True or False value for verbosity. This allows other scripts to assign it to a variable once after determining verbosity level for the script and allows for easy printing of extra information without constantly needing to check verbosity level.

Example usages: vprint = varconlib.verbose_print(True)