Added verbose flag, combined metallicity and temperature plots into one as subplots, general plot improvements.
Plots vs. temperature and metallicity now are subplots on the same image to make it easier to compare the two for a given pair. Files are now output to a single directory 'pairs' in preparation for similar plots for transitions. Plot limits standardized, and all subplots share the (appropriate) x- and y-limits.

Added use of verbose_print from varconlib.