Added absolute magnitude and log(g) attributes, save additional data in HDF5 files.
Temperature, metallicity, absolute magnitude, and log(g) are now all collected in a single data file, and saved in the HDF5 file when a Star object is created as optimization to avoid needing to open the file multiple times in future. Setters were added for all attributes to allow them to be reconstructed from the saved data.