Overhaul of HDF5 data saving, updated docstrings, added apparent magnitude.
Added a current version variable which will be written into HDF5 files
going forward. Currently at 1.0.0.

Added apparent magnitude as a property to be saved.

Globally renamed logG to logg.

Updated docstrings to document newly-added arrays such as ones
pertaining to calibration corrections, and various attributes
which were not formerly documented.

Renamed dumpDataToDisk to saveDataToDisk.

Made it so that the constructed HDF5 filename is save as
self.hdf5file, and used as the default value for saveDataToDisk
allowing easy updating of an HDF5 file with new attributes, for
example.