Changed output, removed units in function, added tolerance argument.
Output is now a dictionary of results rather than a tuple, and
the results returned have been consolidated a bit as well.
The way the function was written, the value of sigma_sys returned
was actually modified one last time before breaking out of the while
loop. Now the entire history of sigma_sys for each iteration is
returned, with the last value being the correct value which makes
chi-squared per DOF get within the tolerance.

Use of units is now not allowed in the function, as that seemed
to cause a strange bug whereby the final (previously incorrect)
value of sigma_sys would be returned for all iterations (though
it obviously wasn't affecting the actual values used, as chi^2
still converged to one). Regardless, the function is now unit-
agnostic to prevent this weird, seemingly-cosmetic, bug.

An argument for tolerance was added, with a default value of
0.001.