Added specialAttribute attribute, incremented version.
self.specialAttributes hold a dictionary (possibly empty) which
contains a few recognized keywords for certain categories which
stars may fall into. Currently there are two keywords:
'is_variable', with the value of the type of variable
'has_companion', with the value of the type of companion

The dictionary is updated from a JSON file that may optionally be
present in the star's directory, named
"{star name}_special_attributes.json". If this file is not present
when the Star class is initialized, specialAttributes will be an
empty dictionary.

This should be backwards-compatible at the moment as this new
attribute is not currently saved, so the version has been raised
from 1.0.0 to 1.1.0.