Overhaul of arrays corrected for stellar parameters.
Major refactor of how pair separations are handled and how
arrays of values corrected for stellar parameters are created
and stored.

New arrays paramsOffsetsArray, paramsErrorsArray, and
paramsCorrectionsArray are now created and stored as part of the
HDF5 file saved. The first two include NaNs where outliers exist
from the fitting model function, which must now be given (and
defaults to "cross_term") in the __init__ method.
paramsCorrectionsArray contains the actual corrections applied to
each transition, for both the pre- and post-fiber change eras.

createPairSeparationsArray and createPairSepErrsArray now draw on
the aforementioned arrays, and can contain NaNs as well.