Added pairSepSysErrorsArray, updated formatPairData
Removed the getPairSeparation method which wasn't being used.

Added a new array of size (2, number of individual pair measurements)
to hold the systematic error for each pair (quadrature sum of the
systematic errors for each component transition which come from
the new array paramsSysErrorsArray. The two rows hold values
for pre- and post-fiber change eras.

Updated formatPairData to include overall systematic error for
the pair and for the component transitions.