Add more ways of checking convergence to a final value.
The change to sigma_sys is now the 2/3 power of the chi^2_nu
value instead of the 1/2 power. Hopefully this keeps the changes
a bit more linear close to 1.

Added a check for whether chi^2_nu itself is converging; it may
be the case that, for a given set of residuals, its value may
never get up to with the tolerance of 1, even if sigma_sys goes
to 0, which would take forever. If chi^2_nu has converged to within
the tolerance for the last 100 iterations, the loop will now break.

Tweaked the change at 500 iterations to have a few more checks; it
might not be necessary any more with the previous addition.

Fixed a typo and removed a now-unnecessary line of code.