Removed references to fit_params, and fixed case of no observations.
Now that outlier rejection happens in the fitting process the code
for dealing with it using pre-determined fitting coefficients is no
longer needed, and has been removed.

Also fixed a case where stars with no observations (either masked out
or simply because none were good) were crashing; this was taken care
of in the previously-mentioned code for outlier rejection, but not in
the code for looking at the "raw" values, which is probably how it
survived so long.