Major overhaul of process for creating transition/pair model-corrected arrays.
Renamed methods to create[Transition|Pair]ModelCorrectedArrays,
made them conditional in __init__ upon finding the appropriate
hdf5 file holding the necessary fitting information, made both
loading and saving a star's data to hdf5 able to handle missing
attributes for easier inter-version compatability.
Both methods can take just a model function name, but can also take
a filename to directly access instead of creating it from the function
name.
Added several attributes related to pairs, and renamed several related
to transitions for a more cohesive and easily-memorable naming scheme.

Overhauled docstrings in a number of cases to bring them up to speed
with the changes.

formatPairData is currently broken, and need overhauling to bring it
in line with the new prcoess.