Update requirements, big refactor
This commit increases the hickle and h5py minimum versions,
and adds p_tqdm as a requirement. It overhauls the process for
initializing stars, partly in an attempt to allow for
concurrency in the process (which failed), but also to make the
code cleaner and quicker to execute, while adding two new
attributes to star.Star, arrays of the calibration source and
file used for calibration for each observation.

This commit also breaks out the logic for reading values from an
observation's pickle files into its own private method.